+++ hookee/conf.py
@@ -1,19 +1,10 @@
 import os
-from types import ModuleType
 
 import confuse
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
 __version__ = "0.0.7"
-
-
-def as_template(value, default=None):
-    param = confuse.as_template(value)
-    param.default = default
-
-    return param
-
 
 template = {
     "auth_token": confuse.String(default=None),
@@ -24,9 +15,8 @@ template = {
     "plugins_dir": confuse.Filename(),
     "plugins": list,
     "console_width": confuse.Integer(default=80),
-    # TODO: these being modules doesn't work outside of, well, modules, and the ideal is you want to just point it to a script relative to where you are on the CLI
-    "last_request": as_template(ModuleType, default=None),
-    "last_response": as_template(ModuleType, default=None),
+    "last_request": confuse.Filename(default=None),
+    "last_response": confuse.Filename(default=None),
 }