+++ hookee/util.py
@@ -1,3 +1,5 @@
+import sys
+
 import click
 
 __author__ = "Alex Laird"
@@ -22,12 +24,16 @@ class PrintUtil:
     def print_config_update(self, msg):
         click.secho("\n--> {}\n".format(msg), fg="green")
 
-    def print_open_header(self, title, fg="green"):
+    def print_open_header(self, title, delimiter="-", fg="green"):
         width = int((self.console_width - len(title)) / 2)
 
         click.echo("")
-        click.secho("{}{}{}".format("-" * width, title, "-" * width), fg=fg, bold=True)
+        click.secho("{}{}{}".format(delimiter * width, title, delimiter * width), fg=fg, bold=True)
         click.echo("")
 
     def print_close_header(self, delimiter="-", fg="green"):
         click.secho(delimiter * self.console_width, fg=fg, bold=True)
+
+
+def is_python_3():
+    return sys.version_info >= (3, 0)