+++ CHANGELOG.md
@@ -6,8 +6,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased](https://github.com/alexdlaird/hookee/compare/0.0.6...HEAD)
 ### Added
 - Additional `ngrok` passthrough args to the CLI: `region`, `subdomain`, and `auth`.
-- `--last_request` as an arg to CLI. Without the need for plugins, `last_request.run(request)` will be called after all plugins when processing a request to the default `/webhook`.
-- `--last_response` as an arg to CLI. Without the need for plugins, `last_response.run(request, response)` will be called after all plugins when processing a request to the default `/webhook`.
+- `--last_request` as an arg to CLI, which can point to a Python script, where `last_request.run(request)` will be called after all plugins have processed a request to the default `/webhook`.").
+- `--last_response` as an arg to CLI, which can point to a Python script, where `last_response.run(request, response)` will be called after all plugins have generated the default `/webhook`'s response..
 
 ### Fixed
 - Stability improvements.